requirements.txt
python-dotenv==1.0.0
redis==4.5.4
sanic==23.3.0
sentence_transformers==2.2.2
openai==0.27.7
qdrant_client==1.3.2
